Add support for multiple rtn. reqs. plus tidy up (#977)
https://eaflood.atlassian.net/browse/WATER-4463

The first problem we had to tackle was how we hold the selections a user is making on _any_ journey in the app till they are ready to commit them.

With that resolved (welcome the `SessionModel`), we built the manual 'returns required' and 'no returns required' journeys for the return requirements feature. That meant saving and then replaying what a user had selected so we could get to the 'Check your answers' page with something for the QA team to look at.

Now we are on the cusp of building the return requirement summary card for 'Check your answers'. But hot on the heels of that will be the need to allow users to add additional return requirements! We ..cough.. haven't catered for that in our session handling and the journey 😰.

The key things we need to be able to do on any page in the journey is

- know which requirement we are extracting a value from for presenting on the page
- know which requirement we are updating when a user submits a new value

This change implements that support.

---

Doing this work meant we needed to touch _a lot_ of the existing files. What started as some minor [boy scout tidying](https://www.oreilly.com/library/view/97-things-every/9780596809515/ch08.html) quickly snowballed into a major tidy-up of the feature.

- We've fixed some missed features like backlinks needing to return users to 'check your answers'
- We've corrected some overlooked changes like 'return requirements -> requirements for returns'
- We've moved logic out of the views and into the presenters
- We've cleaned up some of the tests that were checking things already covered, stubbing things unnecessarily, or not covering branches of logic in the code
- We've tried to make things as consistent as possible across the various presenters, services, validators and views
- We've renamed `/check-your-answers` to just `/check`
- We've renamed `/cancel-requirements` to just `/cancel`

We are aware this is not how we normally like to do things. This PR has become much too large and unfocused. But at some point, we crossed a line so decided to plough on.