Remove two-part tariff check endpoint from project
https://eaflood.atlassian.net/browse/WATER-4057

To support the initial stages of building the new SROC two-part tariff match & allocate engine we created a `/check` endpoint. This would allow devs and testers alike to trigger the engine and review the results.

We are now sufficiently far along in the build of two-part tariff that we can perform the same checks using the review pages we've built. In fact, the focus has been on those pages for so long we'd need to make changes to our `/check` endpoint to get it up to par!

It was only intended to be a temporary thing and now seems a good time to say goodbye.

So, hank you for your services to the cause `/check`! 🫡 😢