Ensure parens used in all arrow functions
https://github.com/DEFRA/water-abstraction-team/issues/115

We have a written convention that [all arrow functions should wrap params in parens even when there is only one](https://github.com/DEFRA/water-abstraction-team/blob/main/coding_conventions.md#arrow-functions).

We think code is clearer to follow and understand when it is _consistent_. JavaScript will allow you to skip adding them when there is only one param to an arrow function. But we think this inconsistency is not helpful to those new to the language or the team.

So, it is our convention to _always_ add parens. Now we are using ESLint we can make the tool highlight when this is missed rather than relying on reviewers.