Use `yar` to store 2PT Review Licences filter data (#999)
https://eaflood.atlassian.net/browse/WATER-4462

As part of adding pagination to the results on the two-part tariff "Review Licences" page. The way the data from the filter is passed back to the view needs to change.

How it is currently set up will mean that once the user selects a different page using the pagination, which triggers a new GET request, the filter data will be lost as it isn't persisted anywhere.

In this PR the filter data is going to be persisted using `yar` to resolve the issue.