Suggested View contact details changes
https://eaflood.atlassian.net/browse/WATER-4433

This relates to [Add View Licence contact details page](#993). What started as experimenting with an alternate fetch implementation has grown into a detailed comparison of the original changes and our current patterns and conventions.

In this we have tried to highlight how we think the changes would be implemented if they followed our current norms.

These are suggested and can be challenged. But we felt in this case "show don't tell" would be more appropriate to support the conversation!