Fix unsent bills displaying in licence bills tab
https://eaflood.atlassian.net/browse/WATER-4316

In [Add View Licence Bills page](https://github.com/DEFRA/water-abstraction-system/pull/986) we added support for displaying the bills linked to a licence in the view licence page's 'Bills' tab.

Only we didn't spot that it should only be 'sent' bills.

This change updates the relevant fetch service to ensure the bills we get back are only those with a status of 'sent'.