Highlight existing test does not link with bill run
The service has a withGraphFetched that we are assuming is finding the matching bill run as we know the presenter this data gets passed to is expecting it.

However, coming in to add a test case for our currently failing scenario we spotted that the existing test is not actually checking for the bill run. Nor was the bill run created actually linked to the bill.

So, we remove everything to do with the bill run from this test to highlight the point we'll need to do some refactoring to ensure the existing scenarios are covered before we make our changes.