Update points page to use point id as the value (#1016)
https://eaflood.atlassian.net/browse/WATER-4295

As part of our work to get the requirements for returns set up we need to know which of the points the user has selected in the manual journey so that we can then get the information later at the check your answers page.

This PR changes the logic to store the ID rather than the string which describes the point. that way we can look it up again.