Add Scheduled Notification view and model
The communication tab on the view licence page need to get the scheduled notification in order o render a list of communications.

Scheduled notifications for the communications tab appear to need the events table data to help render the information with context.

The schedule notification model has been added to handle the interaction with the database.