You are about to remove these return requirements (#996)
* You are about to remove these return requirements

https://eaflood.atlassian.net/browse/WATER-4304

During the return requirements journey, users are given the option to add more requirements to
the returns. As a result, users need to have an option to remove a single requirement at a time.

This PR is focused on setting up the page and functionality for users to remove a single requirement.