Add View Licence communications
https://eaflood.atlassian.net/browse/WATER-4315

The existing service handling view licence is slow because it loads all the data for the tabs in one render. Work has been done previously to refactor the summary page to load only the summary information.

This change will add a communications controller, service and presenter to handle the communications tab data introducing a customer contacts presenter and fetch service.

This will share the same view as the summary page and load the same 'common data' established in previous work.