Alter 2PT Review Licences filter behaviour https://eaflood.atlassian.net/browse/WATER-4462
Currently the filter data that is held in the cookie for the "Review Licences" page will persist each time the user accesses the page, regardless of which 2PT bill run they are reviewing.

This is not desirable behaviour, so in this PR the cookie data will now include the `billRunId` to which it relates so that it is only persisted for the bill run that the user set the filter on.