Alter 2PT Review Licences filter behaviour (#1024)
https://eaflood.atlassian.net/browse/WATER-4462

Currently, the filter data that is held in the cookie for the "Review Licences" page will persist each time the user accesses the page, regardless of which 2PT bill run they are reviewing.

This is not desirable behaviour. So in this PR the identifier for the cookie data will now include the `billRunId` to which it relates so it only persists data for the bill run the user set the filter on.