Fix QA issues found with the review adjustment factors validation (#1027)
https://eaflood.atlassian.net/browse/WATER-4439

During QA it was found that the error message for when a factor greater than 1 is entered was incorrect in the ticket & subsequently on the page. It used to read "factor must be less than 1".  When it should have read "factor must be equal to or less than 1".

Also the ticket originally specified that the Charge Adjustment should be validated to 2DP. This was incorrect, it should be validated to 15DP like the Aggregate factor.

These two issues will be corrected in this ticket.