Add sentence case function & rename capitalize  (#1029)
https://eaflood.atlassian.net/browse/WATER-4315

As part of our work to re-create the communications tab in the view licence page, we needed to display some text from the DB in [Sentence case](https://titlecaseconverter.com/blog/sentence-case-vs-title-case/). At the same time, we also needed to capitalize the first character of a word. The superstar @jonathangoulding solved both of these problems in [their change](https://github.com/DEFRA/water-abstraction-system/pull/1020).

But wait! We already have a `capitalize()` method in the base presenter. We were ready to suggest the change when we realised it wouldn't solve the sentence case issue.

Eeek! We need to keep the sentence case solution, but it relies on the title case method @jonathangoulding added.

This change is to sort things out so we can solve the problems in that change. Our reasoning is that if we have a `sentenceCase()` solution, its ideal home is with similar functions in the base presenter. But it also highlighted that the world refers to what we are currently doing in `capitalize()` as 'title case'.

So, we rename it as part of this change.