Refactor purposes page to use purpose ids as values (#1025)
https://eaflood.atlassian.net/browse/WATER-4294

As part of our work to get the requirements for returns set up we need to know which of the purposes the user has selected in the manual journey so that we can then get the information later at the check your answers page.

This PR changes the logic to store the ID rather than the string which describes the purpose. that way we can look it up again.