Fix view licence summary for incomplete licences (#1033)
https://eaflood.atlassian.net/browse/WATER-4322

In testing, we have found licences that have incomplete information. For example, `28/39/29/0048C` has no contacts, not even a licence holder, nor any licence version records.

The data doesn't appear to exist in the NALD extract we receive, so we have nothing we can use to fill in the missing pieces. But we don't want our licence summary to break when a user searches for the licence and then clicks through.

So, these changes ensure the page _shouldn't_ error when a user selects one of these incomplete licences.