Fix adjustment factors link
https://eaflood.atlassian.net/browse/WATER-4439

During the testing for the above ticket it was noted that if an adjustment factor is removed from a licence on the set adjustment factor page, then the link is removed from the charge reference details page after. This is incorrect as the link should always stay if a charge factor was present initially. This PR adds in the functionally for the link to always be present if a charge factor was initially on the licence.