Calculate and display expected charge during 2PT review https://eaflood.atlassian.net/browse/WATER-4441
Add functionality to the Review Charge Reference page to allow the user to preview the charge for the Charge Reference.

The preview of charge will be triggered by a button on the Review Charge Reference details page, and will display the result in a banner at the top of the screen.