View licence communications remove alert
The implementation of alert was not understood from the old system implementation.

This change removes the idea of an alert and a label and merges them together to simplify the view and presenter.

The outcome should be the label is returned unless the sendingAlertType is Water abstraction alert the it should append ' - Water abstraction alert' to the end of the alert type.

For example - Stop - Water abstraction alert