Content change on set adjustment factors page 2pt
https://eaflood.atlassian.net/browse/WATER-4439

During the testing for the above ticket it was decided that a content change was necessary on the set adjustment factors page. This PR is removing the 'optional' text from the page. The PR is also fixing a small bug, where if a user enters 0 for the adjustment factors (unlikely) that the 0 will appear pre populated on the input boxes if they were to go back to the page again (it currently is not populating the boxes with a 0).