Calculate and display expected charge during 2PT review (#1036)
https://eaflood.atlassian.net/browse/WATER-4441

Add functionality to the Review Charge Reference page to allow the user to preview the charge for the Charge Reference.

The preview of the charge will be triggered by a button on the Review Charge Reference details page and will display the result in a banner at the top of the screen.

This is the first of two PRs. In this PR a new Charging Module "request" is being added to enable a request to be sent to the Charging Module to calculate a charge. The next PR will cover adding the functionality to the screens to enable a charge to be calculated for a charge reference.