View licence set up agreements
This will add an agreement table and buttons to the view licence set up page.

The agreements action are based on the user role and state of an agreement.

An authorised user can delete/end/recalculate an agreement as well as set up a new agreement

This change is part of the charge information page replacement seen in this previous work - https://github.com/DEFRA/water-abstraction-system/pull/1034