Fix user seeding not creating user groups
https://eaflood.atlassian.net/browse/WATER-3981

The [water-abstraction-service](https://gitub.com/DEFRA/water-abstraction-service) has a mechanism for seeding data for testing that relies on loading in YAML fixture files, sort of. Some content comes from these but other stuff is hard-coded into the 'loaders'.

Long story short we're going to replace it with something simpler yet more flexible. We initially were going to use [Knex's seeding functionality](https://knexjs.org/guide/migrations.html#seed-files) to generate the data.

But we've had a change of mind (look out for `/data/load` coming in a future change!) We still need users though and we can simplify things in our acceptance tests if they can assume the different types of users have been created.

That's when we spotted this isn't actually working! 🤦

It is creating the users but not linking them to the groups that give them the permissions needed. This change fixes the user seed.