Two-part tariff review pages housekeeping
During work on one of the final two-part tariff review pages, it was pointed out there are a few inconsistency's in some of the pages code. This PR is for updating all the buttons on the two-part tariff review pages to have the preventDoubleClick attributes set as true. This is done on some pages but not all.