Two-part tariff review pages housekeeping (#1050)
During work on one of the final two-part tariff review pages, it was pointed out there are a few inconsistency's in some of the pages code.
This PR is for updating all the buttons on the two-part tariff review pages to have the preventDoubleClick attributes set as true. This is currently done on some pages but not all.
This PR is also removing an unnecessary 'fromFinanicalYearEnding' property that it being fetched on one of our fetch services. The property isn't used once it's been fetched so it makes sense to remove it.