Calculate and display expected charge during 2PT review - Part 2 (#1046)
https://eaflood.atlassian.net/browse/WATER-4441

Add functionality to the Review Charge Reference page to allow the user to preview the charge for the Charge Reference.

The preview of the charge will be triggered by a button on the Review Charge Reference details page and will display the result in a banner at the top of the screen.

This is the second of two PRs. In the previous PR https://github.com/DEFRA/water-abstraction-system/pull/1036 a new Charging Module "request" was added to enable a request to be sent to the Charging Module to calculate a charge.

This PR will cover adding the functionality to the screens to enable a charge to be calculated and displayed for a charge reference.