Add the returns cycle text to the returns requirements check page (#1075)
* Add the returns cycle text to the returns requirements check page

Due to the estimated complexity of adding fetch services for purposes, points and other display items. We have broken down the complex display items into their own branches of work.

https://eaflood.atlassian.net/browse/WATER-4386

This branch will add returns cycles to the return requirement summary card. As of this commit there is only 'summer' and 'winter and all year' returns cycles

Previous work done here - #1019

* chore: pre pr fixes

* fix: test failing missing returns cycle assert

* chore: alpha order

* Update app/presenters/return-requirements/check/returns-requirements.presenter.js

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

* Update test/presenters/return-requirements/check/returns-requirements.presenter.test.js

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

* fix: use ternary for returns cycle

assume that if the return is not summer it is winter as there are only two options

---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>