Tidy up the file names of legacy migrations (#1077)
This is just some simple housekeeping to keep things in order.

The legacy migrations are purely used when creating the test database. They support our unit tests because we have chosen not to stub out the DB layer.

We went through an exercise of grouping the migrations by schema so that we can more easily see what legacy tables we have migrations for. However, we've let that slide recently.

And, while we are fixing filenames, we're also updating them all to separate the timestamp from the given name with an underscore. This is what would happen if you created the migration the normal way.