Remove blipp plugin https://github.com/DEFRA/water-abstraction-team/issues/111
We and it seems most other Defra services add this as standard to projects.

But we've been (rightly) challenged on what value this adds. As the project progresses the number of routes is going to expand so that routing table is going to get very noisy. It's an extra dependency and we even have untested logic in `server.js` to only add the plugin when in non-prod environments.

Keeping to our ethos of minimal and simple, why do we have this dependency?!

Having discussed it no one could make a case for keeping it. So, we think it is time to say goodbye to **Blipp** 😢👋