Add return requirements models to project (#1071)
https://eaflood.atlassian.net/browse/WATER-4467

So far, the new returns requirements functionality supports setting up a new return version and its requirements. But the data is held only in a temporary session.

Alongside this, we've been making changes to the legacy return requirements tables so they can support the new properties we will record in our version of the returns requirement setup journey.

We now want to create new return requirements by copying an existing one, persist those that have been setup during the journey, and in the future, generate return logs using this information.

All this will require the tables to be represented as [Objection.js models](https://vincit.github.io/objection.js/). This change adds the models, the views, and all the supporting elements in preparation for this work.