Add copy existing rtn req. option to setup journey
https://eaflood.atlassian.net/browse/WATER-4355

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to copy data from an existing return requirement when setting up a new one. However, the option must only appear if one exists to copy.

This change adds the logic to check for existing returns and the option to the "How do you want to set up the returns" page.