Fetch the current return versions in initiate
We're already fetching a bunch of data for the licence during the initiate process. We could have built a service to confirm if return versions exist for a licence. This would then need to be called in the `Setup` and `SubmitSetup` services.

But then we'd need another service to actually fetch details for each version in order to have something to display to the user. Of course, this service could support both functions. But then we're fetching the same data twice, but using it only once.

Fetching and storing it at this point means we make the call once and then it is always available. There is a chance it isn't used, for example the user bails early or selects a different path. But we think this solution is the simplest and easiest to maintain out of the options.