Add copy existing rtn req. option to setup journey (#1081)
https://eaflood.atlassian.net/browse/WATER-4355

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to copy data from an existing return requirement when setting up a new one. However, the option must only appear if one exists to copy.

This change adds the logic to check for existing returns and the option to the "How do you want to set up the returns" page.

---

**Note about where we fetch the return versions**

We're already fetching a bunch of data for the licence during the initiate process. We could have built a service to confirm if return versions exist for a licence. This would then need to be called in the `Setup` and `SubmitSetup` services.

But then we'd need another service to actually fetch details for each version in order to have something to display to the user. Of course, the fetch service could support both functions. But then we're fetching the same data twice, but using it only once.

Fetching and storing it at this point means we make the call once and then it is always available. There is a chance it isn't used, for example, if the user bails early or selects a different path. But we think this solution is the simplest and easiest to maintain out of the options.