Complete Select existing return requirement page (#895)
https://eaflood.atlassian.net/browse/WATER-4283

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to copy data from an existing return requirement when setting up a new one. We [made the option available](https://github.com/DEFRA/water-abstraction-system/pull/1081) in the `/setup` page.

When a user selects this option they'll be taken to the `Select an existing return requirement from` page. This change is about completing that page and the feature.

- add the controls to the page. These will be dynamically generated based on the licences current return versions
- add validation to the page
- retrieve the requirements for the selected return version and populate the session with them

Once the page has completed it's work the user will be redirected to the `/check` page where they can see the results of their action.