Exclude not-required rtn versions from copy exist
https://eaflood.atlassian.net/browse/WATER-4283

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to copy data from an existing return requirement when setting up a new one. We [made the option available](https://github.com/DEFRA/water-abstraction-system/pull/1081) in the `/setup` page.

We then implemented the logic to fetch the existing data and transform it into a something that can be stored in the session for later persisting back as a new record.

One of the new features we're enabling with our takeover is the ability to record where a return requirement is _not_ needed. Our new journey allows for this.

When this happens though, we'll need to persist a return version to record it. It will be 'current' but it will have no return requirements linked to it.

Because of this we need to tweak the query which fetches which return versions can be copied from to exclude those without any return requirements.