Adding tags to view for acceptance tests (returns) (#1070)
* Adding tags to view for acceptance tests (returns)

https://eaflood.atlassian.net/browse/WATER-4485

When writing acceptance tests, due to the way [crypress](https://docs.cypress.io/guides/overview/why-cypress) works, some components of a page can be hard to locate.

For example, imagine a summary list which has nested values. This might be the address on the page provided by cypress:

```
':nth-child(2) > .govuk-summary-list.govuk-\!-margin-bottom-2 > :nth-child(1) > .govuk-summary-list__value'
```

To make it easier to locate the elements on the page we want to test, we add tags like so:
```
'<span data-test="meta-data-reason">' + reason + '</span>'
```

This PR is focused on adding tags to the views in the returns required journey as well as adding more points to the permit licence helper to provide 2 points for the return requirement points page.