Correct copy existing return req. page title
https://eaflood.atlassian.net/browse/WATER-4283

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to copy data from an existing return requirement when setting up a new one. We [made the option available](https://github.com/DEFRA/water-abstraction-system/pull/1081) in the `/setup` page.

We then did the actual work of allowing users to select the return requirements to copy and saving them to the session in [Complete Select existing return requirement page](https://github.com/DEFRA/water-abstraction-system/pull/895).

But we've since learnt that the title we used for the page is not right. This change corrects it.